Added pastie.org and notepad.cc based on changes from @Wingless-Archangel.
